frontend/requirements.txt
Flask==0.10.1
pyyaml==3.12
